+++ bvm/ballerina-runtime/src/test/java/io/ballerina/runtime/test/config/TomlProviderTest.java
@@ -58,7 +58,6 @@ import java.util.Set;
 import java.util.function.Function;
 
 import static io.ballerina.runtime.api.PredefinedTypes.TYPE_ANYDATA;
-import static io.ballerina.runtime.api.PredefinedTypes.TYPE_BOOLEAN;
 import static io.ballerina.runtime.api.PredefinedTypes.TYPE_BYTE;
 import static io.ballerina.runtime.api.PredefinedTypes.TYPE_DECIMAL;
 import static io.ballerina.runtime.api.PredefinedTypes.TYPE_FLOAT;
@@ -136,46 +135,6 @@ public class TomlProviderTest {
                         ROOT_MODULE, "decimalArr", new BIntersectionType(ROOT_MODULE, new BType[]{}, TypeCreator
                         .createArrayType(TYPE_DECIMAL), 0, false), true), decimalArrayGetFunction,
                         expectedDecimalArray}
-        };
-    }
-
-    @Test(dataProvider = "record-data-provider")
-    public void testTomlProviderRecords(String variableName, Map<String, Field> fields,
-                                        Map<String, Object> expectedValues) {
-
-        RecordType type =
-                TypeCreator.createRecordType("Person", ROOT_MODULE, SymbolFlags.READONLY, fields, null, true, 6);
-        VariableKey recordVar = new VariableKey(ROOT_MODULE, variableName, type, true);
-        Map<Module, VariableKey[]> configVarMap = Map.ofEntries(Map.entry(ROOT_MODULE, new VariableKey[]{recordVar}));
-        RuntimeDiagnosticLog diagnosticLog = new RuntimeDiagnosticLog();
-        ConfigResolver configResolver = new ConfigResolver(configVarMap, diagnosticLog,
-                List.of(new TomlFileProvider(ROOT_MODULE, getConfigPath("RecordTypeConfig.toml"),
-                        configVarMap.keySet())));
-        Map<VariableKey, Object> configValueMap = configResolver.resolveConfigs();
-
-        Assert.assertTrue(configValueMap.get(recordVar) instanceof BMap<?, ?>,
-                "Non-map value received for variable : " + variableName);
-        BMap<?, ?> record = (BMap<?, ?>) configValueMap.get(recordVar);
-        for (Map.Entry<String, Object> expectedField : expectedValues.entrySet()) {
-            BString fieldName = fromString(fields.get(expectedField.getKey()).getFieldName());
-            Assert.assertEquals((record.get(fieldName)), expectedField.getValue());
-        }
-    }
-
-    @DataProvider(name = "record-data-provider")
-    public Object[][] recordDataProvider() {
-        Field name = TypeCreator.createField(TYPE_STRING, "name", SymbolFlags.REQUIRED);
-        Field nameReadOnly = TypeCreator.createField(TYPE_STRING, "name", SymbolFlags.READONLY);
-        Field age = TypeCreator.createField(TYPE_INT, "age", SymbolFlags.OPTIONAL);
-        return new Object[][]{
-                {"requiredFieldRecord", Map.ofEntries(Map.entry("name", name)),
-                        Map.ofEntries(Map.entry("name", fromString("John")))},
-                {"readonlyFieldRecord", Map.ofEntries(Map.entry("name", nameReadOnly)),
-                        Map.ofEntries(Map.entry("name", fromString("Jade")))},
-                {"optionalFieldRecord", Map.ofEntries(Map.entry("name", nameReadOnly), Map.entry("age", age)),
-                        Map.ofEntries(Map.entry("name", fromString("Anna")), Map.entry("age", 21L))},
-                {"optionalMissingField", Map.ofEntries(Map.entry("name", nameReadOnly), Map.entry("age", age)),
-                        Map.ofEntries(Map.entry("name", fromString("Peter")))},
         };
     }
 
@@ -801,62 +760,4 @@ public class TomlProviderTest {
             }
         }
     }
-
-    @Test
-    public void testAdditionalFieldsInOpenRecords() {
-        RecordType recordType = TypeCreator.createRecordType("Person", ROOT_MODULE, SymbolFlags.READONLY,
-                new HashMap<>(), TYPE_ANYDATA, false, 6);
-        VariableKey recordVar = new VariableKey(ROOT_MODULE, "recordVar", recordType, true);
-        RuntimeDiagnosticLog diagnosticLog = new RuntimeDiagnosticLog();
-        ConfigResolver configResolver = new ConfigResolver(Map.ofEntries(Map.entry(ROOT_MODULE,
-                new VariableKey[]{recordVar})), diagnosticLog, List.of(new TomlFileProvider(ROOT_MODULE, getConfigPath(
-                "OpenRecordConfig.toml"), Set.of(ROOT_MODULE))));
-        Map<VariableKey, Object> configValueMap = configResolver.resolveConfigs();
-        Assert.assertEquals(diagnosticLog.getErrorCount(), 0);
-        Assert.assertEquals(diagnosticLog.getWarningCount(), 0);
-        Object recordValue = configValueMap.get(recordVar);
-        Assert.assertTrue(recordValue instanceof BMap<?, ?>, "Non-map value received");
-        BMap<?, ?> record = (BMap<?, ?>) recordValue;
-        Assert.assertEquals(record.getIntValue(fromString("intValue")), (Long) 25L);
-        Assert.assertEquals(record.getFloatValue(fromString("floatValue")), 25.54);
-        Assert.assertEquals(record.getStringValue(fromString("stringValue")), fromString("test"));
-        Assert.assertEquals(record.getBooleanValue(fromString("booleanValue")), (Boolean) true);
-    }
-
-    @Test(dataProvider = "record-rest-provider")
-    public void testRestFieldsInRecords(Type restFieldType, Map<String, Object> expectedValues, String varName) {
-        RecordType recordType = TypeCreator.createRecordType("Person", ROOT_MODULE, SymbolFlags.READONLY,
-                new HashMap<>(), restFieldType, false, 6);
-        VariableKey recordVar = new VariableKey(ROOT_MODULE, varName, recordType, true);
-        RuntimeDiagnosticLog diagnosticLog = new RuntimeDiagnosticLog();
-        ConfigResolver configResolver = new ConfigResolver(Map.ofEntries(Map.entry(ROOT_MODULE,
-                new VariableKey[]{recordVar})), diagnosticLog, List.of(new TomlFileProvider(ROOT_MODULE,
-                getConfigPath("RestFieldConfig.toml"), Set.of(ROOT_MODULE))));
-        Map<VariableKey, Object> configValueMap = configResolver.resolveConfigs();
-        Assert.assertEquals(diagnosticLog.getErrorCount(), 0);
-        Object recordValue = configValueMap.get(recordVar);
-        Assert.assertTrue(recordValue instanceof BMap<?, ?>, "Non-map value received");
-        BMap<?, ?> record = (BMap<?, ?>) recordValue;
-        for (Map.Entry<String, Object> expectedField : expectedValues.entrySet()) {
-            BString fieldName = fromString(expectedField.getKey());
-            Assert.assertEquals((record.get(fieldName)), expectedField.getValue());
-        }
-    }
-
-    @DataProvider(name = "record-rest-provider")
-    public Object[][] recordRestDataProvider() {
-        BMap<BString, Object> mapValue = ValueCreator.createMapValue();
-        mapValue.put(fromString("m"), fromString("value"));
-        mapValue.put(fromString("n"), 608L);
-        return new Object[][]{
-                {TYPE_INT, Map.ofEntries(Map.entry("id", 11L), Map.entry("age", 26L)), "intRest"},
-                {TYPE_FLOAT, Map.ofEntries(Map.entry("height", 161.5), Map.entry("weight", 62.3)), "floatRest"},
-                {TYPE_STRING, Map.ofEntries(Map.entry("name", fromString("Hinduja")), Map.entry("city", fromString(
-                        "Colombo"))), "stringRest"},
-                {TYPE_BOOLEAN, Map.ofEntries(Map.entry("isBoolean", true), Map.entry("isByte", false)), "booleanRest"},
-                {TYPE_ANYDATA, Map.ofEntries(Map.entry("intVal", 100L), Map.entry("floatVal", 103.507),
-                        Map.entry("stringVal", fromString("string")), Map.entry("mapVal", mapValue)), "anydataRest"},
-        };
-    }
-
 }
