+++ tests/jballerina-integration-test/src/test/resources/configurables/configStructuredTypesProject/Config_tables.toml
@@ -1,0 +1,211 @@
+[[engineerTable]]
+name="hinduja"
+id=100
+
+[[engineerTable]]
+name="riyafa"
+id=105
+
+[[studentTable]]
+name="manu"
+id=100
+
+[[studentTable]]
+name="riyafa"
+id=105
+
+[[employeeTable]]
+name="hinduja"
+id=102
+
+[[employeeTable]]
+name="manu"
+id=100
+
+[[employeeTable1]]
+name="waruna"
+id=2
+
+[[employeeTable1]]
+name="manu"
+id=7
+
+[[officerTable]]
+name="hinduja"
+id=102
+
+[[officerTable]]
+name="manu"
+id=100
+
+[[officerTable1]]
+name="waruna"
+id=4
+
+[[officerTable1]]
+name="gabilan"
+id=5
+
+[[managerTable]]
+name="gabilan"
+id=101
+
+[[managerTable]]
+name="riyafa"
+id=102
+
+[[teacherTable]]
+name="manu"
+id=77
+
+[[teacherTable]]
+name="riyafa"
+id=88
+
+[[farmerTable]]
+name="waruna"
+id=444
+
+[[farmerTable]]
+name="hinduja"
+id=888
+
+[[personTable]]
+name = "riyafa"
+id = 13
+address.city="Canberra"
+address.country.name="Australia"
+
+[[personTable]]
+name = "gabilan"
+id = 14
+address.city="Paris"
+address.country.name="France"
+
+[[lecturerTable]]
+name="manu"
+department1.name="IT"
+department2.name="Sales"
+department3.name="HR"
+
+[[lecturerTable]]
+name="waruna"
+department1.name="IT"
+department3.name="HR"
+
+[[lawyerTable]]
+name="gabilan"
+address1.city="Colombo"
+address2.city="Kandy"
+address3.city="Galle"
+
+[[lawyerTable]]
+name="riyafa"
+address1.city="Matara"
+address3.city="Gampaha"
+
+[[lecturerTable2]]
+name="manu"
+department1.name="IT"
+department2.name="Sales"
+department3.name="HR"
+
+[[lecturerTable2]]
+name="waruna"
+department1.name="IT"
+department3.name="HR"
+
+[[lawyerTable2]]
+name="gabilan"
+place1.city="Colombo"
+place2.city="Kandy"
+place3.city="Galle"
+
+[[lawyerTable2]]
+name="riyafa"
+place1.city="Matara"
+place3.city="Jaffna"
+
+[[configStructuredTypes.imported_records.doctorTable]]
+name="hinduja"
+id=100
+
+[[configStructuredTypes.imported_records.doctorTable]]
+name="riyafa"
+id=105
+
+[[configStructuredTypes.imported_records.studentTable]]
+name="manu"
+id=100
+
+[[configStructuredTypes.imported_records.studentTable]]
+name="riyafa"
+id=105
+
+[[configStructuredTypes.imported_records.employeeTable]]
+name="hinduja"
+id=102
+
+[[configStructuredTypes.imported_records.employeeTable]]
+name="manu"
+id=100
+
+[[configStructuredTypes.imported_records.employeeTable1]]
+name="gabilan"
+id=2
+
+[[configStructuredTypes.imported_records.employeeTable1]]
+name="riyafa"
+id=3
+
+[[configStructuredTypes.imported_records.officerTable]]
+name="hinduja"
+id=102
+
+[[configStructuredTypes.imported_records.officerTable]]
+name="manu"
+id=100
+
+[[configStructuredTypes.imported_records.officerTable1]]
+name="hinduja"
+id=7
+
+[[configStructuredTypes.imported_records.officerTable1]]
+name="waruna"
+id=8
+
+[[configStructuredTypes.imported_records.managerTable]]
+name="gabilan"
+id=101
+
+[[configStructuredTypes.imported_records.managerTable]]
+name="riyafa"
+id=102
+
+[[configStructuredTypes.imported_records.teacherTable]]
+name="gabilan"
+id=66
+
+[[configStructuredTypes.imported_records.teacherTable]]
+name="riyafa"
+id=77
+
+[[configStructuredTypes.imported_records.farmerTable]]
+name="riyafa"
+id=555
+
+[[configStructuredTypes.imported_records.farmerTable]]
+name="hinduja"
+id=666
+
+[[configStructuredTypes.imported_records.personTable]]
+name = "gabilan"
+id = 133
+address.city="Berlin"
+address.country.name="Germany"
+
+[[configStructuredTypes.imported_records.personTable]]
+name = "riyafa"
+id = 144
+address.city="Madrid"
+address.country.name="Spain"
