+++ tests/jballerina-integration-test/src/test/resources/configurables/configStructuredTypesProject/configStructuredTypes/modules/imported_records/record_map_test.bal
@@ -14,47 +14,10 @@
 // specific language governing permissions and limitations
 // under the License.
 
-import configStructuredTypes.mod1;
+import configStructuredTypes.type_defs;
 import testOrg/configLib.mod1 as configLib;
 import ballerina/test;
 
-public type Doctor record {|
-    string name = "";
-    int id = 555;
-|};
-
-type Department readonly & record {|
-    string name;
-|};
-
-type Subject map<string>;
-
-configurable Doctor & readonly doctor = ?;
-configurable configLib:Manager & readonly manager = ?;
-configurable configLib:Teacher & readonly teacher = ?;
-configurable configLib:Farmer farmer = ?;
-configurable mod1:Student & readonly student = ?;
-configurable mod1:Officer officer = ?;
-configurable mod1:Employee employee = ?;
-
-configurable table<Doctor> & readonly doctorTable = ?;
-configurable table<configLib:Manager> & readonly managerTable = ?;
-configurable table<configLib:Teacher> & readonly teacherTable = ?;
-configurable table<configLib:Farmer> & readonly farmerTable = ?;
-configurable table<mod1:Student> & readonly studentTable = ?;
-configurable table<mod1:Officer> & readonly officerTable = ?;
-configurable table<mod1:Employee> & readonly employeeTable = ?;
-
-configurable mod1:Employee & readonly employee1 = ?;
-configurable table<mod1:Officer & readonly> & readonly officerTable1 = ?;
-configurable table<mod1:Employee & readonly> & readonly employeeTable1 = ?;
-
-// Complex records
-configurable mod1:Person person = ?;
-configurable table<mod1:Person> & readonly personTable = ?;
-configurable mod1:Person[] & readonly personArray = ?;
-configurable (mod1:Person & readonly)[] & readonly personArray1 = ?;
-configurable mod1:PersonArray[] & readonly personArray2 = ?;
 
 // Maps
 configurable map<string> & readonly user = ?;
@@ -88,11 +51,11 @@ configurable table<map<Doctor>> & readonly doctorMapTable = ?;
 configurable table<map<Doctor & readonly> & readonly> & readonly readonlyDoctorMapTable = ?; 
 
 // Maps from non-default module
-configurable mod1:IntMap & readonly intMap = ?;
-configurable mod1:StudentMap & readonly studentMap = ?;
-configurable map<mod1:Student & readonly> & readonly readonlyStudentMap = ?;
-configurable table<mod1:StudentMap> & readonly studentMapTable = ?; 
-configurable table<map<mod1:Student & readonly>> & readonly readonlyStudentMapTable = ?;
+configurable type_defs:IntMap & readonly intMap = ?;
+configurable type_defs:StudentMap & readonly studentMap = ?;
+configurable map<type_defs:Student & readonly> & readonly readonlyStudentMap = ?;
+configurable table<type_defs:StudentMap> & readonly studentMapTable = ?; 
+configurable table<map<type_defs:Student & readonly>> & readonly readonlyStudentMapTable = ?;
 
 // Maps from imported module
 configurable configLib:IntMap & readonly libIntMap = ?;
@@ -101,55 +64,6 @@ configurable map<configLib:Manager & readonly> & readonly readonlyManagerMap = ?
 configurable table<configLib:ManagerMap> & readonly managerMapTable = ?; 
 configurable table<map<configLib:Manager & readonly>> & readonly readonlyManagerMapTable = ?;
 
-public function testRecords() {
-    test:assertEquals(doctor.name, "waruna");
-    test:assertEquals(doctor.id, 555);
-    test:assertEquals(student.name, "riyafa");
-    test:assertEquals(student.id, 444);
-    test:assertEquals(employee.name, "manu");
-    test:assertEquals(employee.id, 101);
-    test:assertEquals(employee1.name, "waruna");
-    test:assertEquals(employee1.id, 404);
-    test:assertEquals(officer.name, "gabilan");
-    test:assertEquals(officer.id, 101);
-    test:assertEquals(manager.name, "hinduja");
-    test:assertEquals(manager.id, 107);
-    test:assertEquals(teacher.name, "hinduja");
-    test:assertEquals(teacher.id, 11);
-    test:assertEquals(farmer.name, "manu");
-    test:assertEquals(farmer.id, 22);
-    test:assertEquals(person.name, "hinduja");
-    test:assertEquals(person.id, 100);
-    test:assertEquals(person.address.city, "Kandy");
-    test:assertEquals(person.address.country.name, "Sri Lanka");
-}
-
-public function testTables() {
-    test:assertEquals(doctorTable.toString(), "[{\"name\":\"hinduja\",\"id\":100},{\"name\":\"riyafa\",\"id\":105}]");
-    test:assertEquals(studentTable.toString(), "[{\"name\":\"manu\",\"id\":100},{\"name\":\"riyafa\",\"id\":105}]");
-    test:assertEquals(employeeTable.toString(), "[{\"name\":\"hinduja\",\"id\":102},{\"name\":\"manu\",\"id\":100}]");
-    test:assertEquals(employeeTable1.toString(), "[{\"name\":\"gabilan\",\"id\":2},{\"name\":\"riyafa\",\"id\":3}]");
-    test:assertEquals(officerTable.toString(), "[{\"name\":\"hinduja\",\"id\":102},{\"name\":\"manu\",\"id\":100}]");
-    test:assertEquals(officerTable1.toString(), "[{\"name\":\"hinduja\",\"id\":7},{\"name\":\"waruna\",\"id\":8}]");
-    test:assertEquals(managerTable.toString(), "[{\"name\":\"gabilan\",\"id\":101},{\"name\":\"riyafa\",\"id\":102}]");
-    test:assertEquals(teacherTable.toString(), "[{\"name\":\"gabilan\",\"id\":66},{\"name\":\"riyafa\",\"id\":77}]");
-    test:assertEquals(farmerTable.toString(), "[{\"name\":\"riyafa\",\"id\":555},{\"name\":\"hinduja\",\"id\":666}]");
-    test:assertEquals(personTable.toString(), "[{\"name\":\"gabilan\",\"id\":133," +
-        "\"address\":{\"city\":\"Berlin\",\"country\":{\"name\":\"Germany\"}}},{\"name\":\"riyafa\"," +
-        "\"id\":144,\"address\":{\"city\":\"Madrid\",\"country\":{\"name\":\"Spain\"}}}]");
-}
-
-public function testArrays() {
-    test:assertEquals(personArray.toString(), "[{\"address\":{\"city\":\"Abu Dhabi\",\"country\":{\"name\":\"UAE\"}}," +
-        "\"name\":\"waruna\",\"id\":111},{\"address\":{\"city\":\"Mumbai\",\"country\":{\"name\":\"India\"}}," +
-        "\"name\":\"manu\",\"id\":122}]");
-    test:assertEquals(personArray1.toString(), "[{\"address\":{\"city\":\"New York\",\"country\":{\"name\":\"USA\"}}," +
-        "\"name\":\"manu\",\"id\":800},{\"address\":{\"city\":\"London\",\"country\":{\"name\":\"UK\"}}," +
-        "\"name\":\"hinduja\",\"id\":801}]");
-    test:assertEquals(personArray2.toString(), "[{\"address\":{\"city\":\"New York\",\"country\":{\"name\":\"USA\"}}," +
-        "\"name\":\"riyafa\",\"id\":1000},{\"address\":{\"city\":\"London\",\"country\":{\"name\":\"UK\"}}," +
-        "\"name\":\"waruna\",\"id\":1001}]");
-}
 
 public function testMaps() {
     test:assertEquals(user.toString(), "{\"name\":\"Hinduja\",\"occupation\":\"Software Engineer\","
